* unix/Makefile.in: * unix/tclAppInit.c: Minor cleanup related to Xt notifier.
* unix/tclUnixInit.c (TclpSetInitialEncodings): Tcl now looks for
an encoding subfield in the LANG/LC_ALL variables in cases where
the locale is not found in the locale table.  Ensure that
setlocale() is called at least once so X11 will initialize
properly.  Also, forces the LC_NUMERIC locale to be "C" so numeric
processing in scripts is not affected by the current locale
setting. [Bug: 1989]