* tests/grid.test: Added a test for the consecutive ^ and multiple widget case (bug #1386).
* generic/tkGrid.c: Fixed interpretation of consecutive ^
characters in grid command.  Previously, ^ ^ was interpreted as
meaning that there must be a 2-column widget above to extend,
neglecting the case where there was actually 2 1-column widgets
above.  Now, ^ ^ is interpreted as a possible width; the gridder
will consume as many ^'s as there are columns in the widget, and
leave the rest for the extension of other widgets.  (bug #1386).