* msgcat.n: Added docs for new behavior from patch in [Bug: 4158].
* msgcat.test: Added tests for new behavior from patch in [Bug:
4158].

* msgcat.tcl: Applied patch from [Bug: 4158], which enables
msgcat::mc to search the entire namespace ancestry chain for
message translations (ie, first it checks the current namespace,
then the parent, then the parent's parent, etc).  Also allows the
specification of additional args for msgcat::mc; if extra args are
given, the [format] command is used to substitute the additional
args in the translated message.