<rdar://problem/6744971>: fix file event handler race
Fix subtle race between the point when (due to select() triggering for a
watched descriptor) the tcl notifier thread updates a thread's
readyMasks & signals the thread's tcl runloop source, and the point when
that runloop source's handler actually runs and queues tcl file events
for those readyMasks (if the filehandler in question doesn't already
have a pending file event).

If a file event handler for the same descriptor (from a previous
invocation of the runloop source handler) happens to run between those
points, an extraneous tcl file event can be queued by the runloop source
handler.

FileHandlerEventProc() now clears the current filedescriptor's
readyMasks with the mask being processed, which ensures that invocations
of the tcl runloop source handler run after this point do not queue
events for readyMasks that have already been processed. No events can be
lost by this change since the mask being cleared is that for which an
event handler is about to be run, which will clear the select() trigger
for that mask anyway.